This patch makes generating releases from git a very simple process; you simply edit the makefile's "VERSION" line to the new version, commit that as its own commit, and do: "make test-archive".  That'll make a file in the current directory gnu-efi-$VERSION.tar.bz2 , with its top level directory gnu-efi-$VERSION/ and the source tree under that.
Once you've tested that and you're sure it's what you want to release,
you do "make archive", which will tag a release in git and generate a
final tarball from it.  You then push to the archive, being sure to
include the tag:

git push origin master:master --tags

And upload the archive wherever it's supposed to go.

Signed-off-by: Peter Jones <pjones@redhat.com>
Signed-off-by: Nigel Croxon <nigel.croxon@hp.com>