Fix FTCBFS by exporting CROSS_COMPILE
gnu-efi is in Build-Depends of systemd. As such it is considered part of
the transitive build closure of the essential system and thus needs to
support cross compilation. The package doesn't do that out of the box,
as it always uses "gcc" without any architecture prefix. I am attaching
a patch that exports the CROSS_COMPILE variable used by gnu-efi's build
system when cross compilation is detected and confirm that this patch is
sufficient to support cross building. Please consider applying it.

Closes: #805400